+++ src/structs/id_name_url.cr
@@ -6,7 +6,7 @@ module XIVAPI
     struct IDNameUrl
       # define a JSON mapping to create instances of this struct
       JSON.mapping(
-        id: {type: Int32?, key: "ID"},
+        id: {type: UInt64?, key: "ID"},
         name: {type: String?, key: "Name"},
         url: {type: String?, key: "Url"},
       )
