+++ src/structs/search_result.cr
@@ -7,7 +7,7 @@ module XIVAPI
     struct SearchResult
       # define a JSON mapping to create instances of this struct
       JSON.mapping(
-        id: {type: Int32, key: "ID"},
+        id: {type: UInt64, key: "ID"},
         icon: {type: String, key: "Icon"},
         name: {type: String, key: "Name"},
         url: {type: String, key: "Url"},
